requirements.txt
# Flask WebSocket Server Dependencies
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Audio Processing Dependencies
sounddevice>=0.4.6
numpy>=1.26.0

# WebSocket Client Dependencies
python-socketio>=5.9.0
websocket-client>=1.6.0

# Speech-to-Text Dependencies
SpeechRecognition>=3.10.0
pydub>=0.25.1
openai-whisper>=20231117
soundfile>=0.12.1

# Additional useful packages
python-dotenv>=1.0.0
